Adding special conditions to the interactive module
The current code always assume there is a 'b:interactive' variable set on
the environment. This is not the case when we are using the statement

'b:interactive.is_close_immediately = 1'

It did complain at some point that the variable was non-existing, I added
an if statement that checked if the variable was there, before doing
anything.